requirements.txt
Flask==3.0.3
joblib==1.4.2
matplotlib==3.8.4
mlflow==2.16.2
mlflow_skinny==2.16.2
numpy==2.1.3
pandas==2.2.3
python_dateutil==2.9.0
Requests==2.32.3
scikit_learn==1.4.2
seaborn==0.13.2
statsmodels==0.14.2
streamlit==1.40.0
yfinance==0.2.44
